Moved around code
For more readablility
