+++ app/src/main/res/layout/activity_login_register.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:background="@color/white"
     android:orientation="vertical"
-    tools:context=".LoginRegister">
+    tools:context=".loginregister.LoginRegister">
 
     <com.google.android.material.appbar.MaterialToolbar
         android:id="@+id/toolBar"
