+++ app/src/main/res/layout/activity_tweet_detail.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/slight_light_blue"
-    tools:context=".TweetDetail">
+    tools:context=".feed.TweetDetail">
 
     <com.google.android.material.appbar.AppBarLayout
         android:id="@+id/topAppBar"
